// workspaces/ballerina/type-diagram/src/resources/types.ts
/**
 * Copyright (c) 2022, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
 *
 * This software is the property of WSO2 LLC. and its suppliers, if any.
 * Dissemination of any information or reproduction of any material contained
 * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
 * You may not alter or remove any copyright or other notice from copies of this content.
 */

export enum Colors {
    CELL_DIAGRAM_BACKGROUND = '#eeeffb',
    CONSOLE_CELL_DIAGRAM_BACKGROUND = '#f7f8fb',
    DEFAULT_TEXT = '#40404B',
    DIAGRAM_BACKGROUND = '#FFF',
    GATEWAY = '#c9c9c9',
    PRIMARY = '#5567D5',
    PRIMARY_LIGHT = '#A6B3FF',
    PRIMARY_SELECTED = '#ffaf4d',
    SECONDARY = '#F0F1FB',
    SECONDARY_SELECTED = '#fffaf2',
    SHADED_SELECTED = '#faead2',
    PRIMARY_FOCUSED = '#d59155',
}